refactor(dashboard): wip
